refactor: split option classes and custom account combobox into focused files
