+++ src/components/BankTransactions/RecordManualTransaction/RecordTransactionForm.tsx
@@ -6,7 +6,8 @@ import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
 import { Form } from '@ui/Form/Form'
 import { RecordTransactionCounterpartySelector } from '@components/BankTransactions/RecordManualTransaction/RecordTransactionCounterpartySelector'
 import { type RecordTransactionFormApi, type RecordTransactionVariant } from '@components/BankTransactions/RecordManualTransaction/useRecordTransactionForm'
-import { CustomAccountComboBox, isNewAccountOption } from '@components/CustomAccountComboBox/CustomAccountComboBox'
+import { CustomAccountComboBox } from '@components/CustomAccountComboBox/CustomAccountComboBox'
+import { isNewAccountOption } from '@components/CustomAccountComboBox/utils'
 import { LedgerAccountCombobox } from '@components/LedgerAccountCombobox/LedgerAccountCombobox'
 import { ErrorText } from '@components/Typography/ErrorText'
