+++ src/components/CustomAccountComboBox/CustomAccountComboBox.tsx
@@ -1,64 +1,18 @@
 import { useCallback, useMemo } from 'react'
 import classNames from 'classnames'
-import type { TFunction } from 'i18next'
-import { Check } from 'lucide-react'
 import { useTranslation } from 'react-i18next'
 
 import type { CustomAccount } from '@schemas/customAccounts'
-import { humanizeEnum } from '@utils/format'
 import { useCustomAccounts } from '@hooks/api/businesses/[business-id]/custom-accounts/useCustomAccounts'
-import { COMBO_BOX_CLASS_NAMES } from '@ui/ComboBox/classnames'
 import { CreatableComboBox } from '@ui/ComboBox/CreatableComboBox'
-import { HStack, VStack } from '@ui/Stack/Stack'
-import { Label, Span } from '@ui/Typography/Text'
+import { VStack } from '@ui/Stack/Stack'
+import { Label } from '@ui/Typography/Text'
+import { type AccountOption } from '@components/CustomAccountComboBox/AccountOption'
+import { AccountOptionSlot, AccountSingleValueSlot } from '@components/CustomAccountComboBox/AccountOptionSlots'
+import { formatCreateLabel, isNewAccountOption, NEW_ACCOUNT_VALUE } from '@components/CustomAccountComboBox/utils'
 import { CustomAccountForm } from '@components/CustomAccountForm/CustomAccountForm'
 
 import './customAccountComboBox.scss'
-
-export type AccountOption = {
-  value: string
-  label: string
-  account: Partial<CustomAccount> & Pick<CustomAccount, 'accountName'>
-  __isNew__?: true
-}
-
-const NEW_ACCOUNT_VALUE = 'new_account'
-
-export const isNewAccountOption = (account: AccountOption | null): boolean =>
-  account?.value === NEW_ACCOUNT_VALUE
-
-const formatCreateLabel = (inputValue: string, t: TFunction) =>
-  inputValue
-    ? t('upload:action.create_input_value', 'Create "{{inputValue}}"', { inputValue })
-    : t('upload:action.create_account', 'Create account')
-
-const AccountOptionSlot = ({ option, fallback }: { option: AccountOption, fallback: React.ReactNode }) => {
-  if (option.account && !option.__isNew__) {
-    return (
-      <HStack gap='xs' align='center'>
-        <Check size={16} className={COMBO_BOX_CLASS_NAMES.OPTION_CHECK_ICON} />
-        <VStack>
-          <Span ellipsis>{option.account.accountName}</Span>
-          <Span size='sm' variant='subtle' noWrap>
-            {option.account.institutionName}
-            {' · '}
-            {humanizeEnum(option.account.accountSubtype!)}
-          </Span>
-        </VStack>
-      </HStack>
-    )
-  }
-
-  return fallback
-}
-
-const AccountSingleValueSlot = ({ option, fallback }: { option: AccountOption, fallback: React.ReactNode }) => {
-  if (option.account && !option.__isNew__) {
-    return <Span ellipsis>{option.account.accountName}</Span>
-  }
-
-  return fallback
-}
 
 type CustomAccountComboBoxProps = {
   label: string
