+++ src/components/CustomerSelector/CustomerSelector.tsx
@@ -10,32 +10,9 @@ import { useDebouncedSearchInput } from '@hooks/utils/debouncing/useDebouncedSea
 import { MaybeCreatableComboBox } from '@ui/ComboBox/MaybeCreatableComboBox'
 import { VStack } from '@ui/Stack/Stack'
 import { Label, P } from '@ui/Typography/Text'
+import { CustomerAsOption } from '@components/CustomerSelector/CustomerAsOption'
 
 import './customerSelector.scss'
-
-class CustomerAsOption {
-  private internalCustomer: Customer
-
-  constructor(customer: Customer) {
-    this.internalCustomer = customer
-  }
-
-  get original() {
-    return this.internalCustomer
-  }
-
-  get label() {
-    return getCustomerName(this.internalCustomer)
-  }
-
-  get id() {
-    return this.internalCustomer.id
-  }
-
-  get value() {
-    return this.internalCustomer.id
-  }
-}
 
 type CustomerSelectorBaseProps = {
   selectedCustomer: Customer | null
