+++ src/components/VendorSelector/VendorSelector.tsx
@@ -8,35 +8,9 @@ import { useDebouncedSearchInput } from '@hooks/utils/debouncing/useDebouncedSea
 import { ComboBox } from '@ui/ComboBox/ComboBox'
 import { VStack } from '@ui/Stack/Stack'
 import { Label, P } from '@ui/Typography/Text'
+import { getVendorName, VendorAsOption } from '@components/VendorSelector/VendorAsOption'
 
 import './vendorSelector.scss'
-
-const getVendorName = (vendor: Vendor) =>
-  vendor.individualName ?? vendor.companyName ?? vendor.externalId ?? 'Unknown vendor'
-
-class VendorAsOption {
-  private internalVendor: Vendor
-
-  constructor(vendor: Vendor) {
-    this.internalVendor = vendor
-  }
-
-  get original() {
-    return this.internalVendor
-  }
-
-  get label() {
-    return getVendorName(this.internalVendor)
-  }
-
-  get id() {
-    return this.internalVendor.id
-  }
-
-  get value() {
-    return this.internalVendor.id
-  }
-}
 
 type VendorSelectorProps = {
   selectedVendor: Vendor | null
